+++ bin/src/builder/icons_builder.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:analyzer/dart/element/element.dart';
+import 'package:path/path.dart';
 
 import '../utils/cli_log.dart';
 import '../utils/path_resolver.dart';
@@ -31,16 +32,58 @@ class IconsBuilder extends SpecBuilder {
         if (library != null) {
           for (final element in library.topLevelElements) {
             if (element is PropertyAccessorElement) {
-              final returnElement = element.returnType.element;
-              if (returnElement is ClassElement) {
-                for (final fieldElement in returnElement.fields) {
-                  if (fieldElement.isConst) {
-                    final fieldTypeName = fieldElement.type.getDisplayString(withNullability: false);
-                    if (fieldTypeName == "IconData") {
-                      registrations.write(_buildRegisterIconCall(returnElement, fieldElement));
+              final returnType = element.returnType;
+              if (returnType.toString() != "InvalidType") {
+                // has a known return type
+                final classElements = <ClassElement>[];
+                if (element.name == "icons") {
+                  // found a list of icons
+                  final icons = element.variable.computeConstantValue()?.toListValue();
+                  if (icons != null) {
+                    for (final icon in icons) {
+                      final variable = icon.variable;
+                      final enclosingElement = variable?.enclosingElement;
+                      if (variable != null && enclosingElement != null) {
+                        final iconType = icon.type?.getDisplayString(withNullability: false);
+                        if (iconType == "IconData") {
+                          registrations.write(_buildRegisterIconCall(enclosingElement.displayName, variable.displayName));
+                        } else {
+                          CliLog.warn("Skipped icon. Expected icon item to be of type 'IconData', but found '$iconType'");
+                        }
+                      } else {
+                        CliLog.warn("Skipped icon. Expected icon item to be a member of a class e.i. 'Icons.add' or "
+                            "something similar, but found '$icon'");
+                      }
+                    }
+                  }
+                } else if (element.name == "iconSets") {
+                  // found a list of icon sets
+                  final iconSets = element.variable.computeConstantValue()?.toListValue();
+                  if (iconSets != null) {
+                    for (final iconSet in iconSets) {
+                      final element = iconSet.toTypeValue()?.element;
+                      if (element is ClassElement) classElements.add(element);
+                    }
+                  }
+                } else {
+                  // found individual icon set const
+                  final returnElement = returnType.element;
+                  if (returnElement is ClassElement) classElements.add(returnElement);
+                }
+
+                // process all class elements
+                for (final classElement in classElements) {
+                  for (final fieldElement in classElement.fields) {
+                    if (fieldElement.isConst) {
+                      final fieldTypeName = fieldElement.type.getDisplayString(withNullability: false);
+                      if (fieldTypeName == "IconData") {
+                        registrations.write(_buildRegisterIconCall(classElement.displayName, fieldElement.displayName));
+                      }
                     }
                   }
                 }
+              } else {
+                CliLog.error("InvalidType for property '$element' in '${basename(path)}'");
               }
             }
           }
@@ -62,6 +105,9 @@ class IconsBuilder extends SpecBuilder {
       } else {
         CliLog.success("Skipping icons. No icons found in sources.");
       }
+
+      result.errors = CliLog.errors;
+      result.warnings = CliLog.warnings;
     }
     return result;
   }
@@ -74,8 +120,8 @@ class IconsBuilder extends SpecBuilder {
     return code.toString();
   }
 
-  String _buildRegisterIconCall(ClassElement classElement, FieldElement fieldElement) {
-    final key = "${classElement.displayName}.${fieldElement.displayName}";
+  String _buildRegisterIconCall(String className, String fieldName) {
+    final key = "${className}.${fieldName}";
     final value = key;
     return "  XWidget.registerIcon('$key', $value);\n";
   }