+++ example/lib/xwidget/icon_spec.dart
@@ -2,4 +2,17 @@ import 'package:flutter/material.dart';
 
 // ignore_for_file: unused_element
 
-const Icons? _icons = null;
+// List required icons individually to reduce app size.
+
+const icons = [
+  Icons.add,
+];
+
+// You can also include the entire icon set by simply referencing the
+// enclosing class. This assumes that each icon is a declared as a static
+// field of type IconData.
+
+// const iconSets = [
+//   Icons,
+//   CupertinoIcons,
+// ];