+++ example/lib/xwidget/inflater_spec.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-// Define constants for all widgets and objects you want to generate inflaters for.
+// Add types for all widgets and objects that you want to generate inflaters for.
 // Used by the builder to generate inflaters.
 
 // Best Practice: Keep declarations in alphabetical order. It makes it much easier to
@@ -8,13 +8,15 @@ import 'package:flutter/material.dart';
 
 // ignore_for_file: unused_element
 
-const AppBar? _appBar = null;
-const Center? _center = null;
-const Column? _column = null;
-const FloatingActionButton? _floatingActionButton = null;
-const Icon? _icon = null;
-const MaterialApp? _materialApp = null;
-const Scaffold? _scaffold = null;
-const Text? _text = null;
-const TextStyle? _textStyle = null;
-const ThemeData? _themeData = null;
+const inflaters = [
+  AppBar,
+  Center,
+  Column,
+  FloatingActionButton,
+  Icon,
+  MaterialApp,
+  Scaffold,
+  Text,
+  TextStyle,
+  ThemeData,
+];