+++ test/bin/builder/inflater_builder_test.dart
@@ -5,12 +5,25 @@ import '../../../bin/src/builder/inflater_builder.dart';
 
 
 void main() {
+
   test('Test InflaterBuilder dependency resolution', () async {
     final config = BuilderConfig();
     config.inflaterConfig.sources.add("xwidget|test/fixtures/src/inflater_spec.dart");
     config.inflaterConfig.target = "build/test.dart";
     config.schemaConfig.template = "xwidget|res/schema_template.xsd";
     config.schemaConfig.target = "build/test.xsd";
+
+    final builder = InflaterBuilder(config);
+    final result = await builder.build();
+    expect(result.outputs.length, 2);
+  });
+
+  test('Test new inflater spec format', () async {
+    final config = BuilderConfig();
+    config.inflaterConfig.sources.add("xwidget|test/fixtures/src/inflater_spec.dart");
+    config.inflaterConfig.target = "build/inflaters_test.g.dart";
+    config.schemaConfig.template = "xwidget|res/schema_template.xsd";
+    config.schemaConfig.target = "build/schema_test.xsd";
 
     final builder = InflaterBuilder(config);
     final result = await builder.build();