+++ test/fixtures/src/icons_spec.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+// ignore_for_file: unused_element
+
+const CupertinoIcons? _cupertinoIcons = null;
+
+const icons = [
+  Icons.add,
+  Icons.delete,
+  IconData(0xe758, fontFamily: 'MaterialIcons')
+];
+
+const iconSets = [
+  Icons,
+];